Game

+++ src/main/java/com/example/proiectis/websocket/Client.java
@@ -6,11 +6,11 @@ import org.springframework.web.socket.WebSocketSession;
 @Getter
 public class Client {
 
-    private final int id;
+    private final Long id;
     private final Channel channel;
     private final WebSocketSession session;
 
-    public Client(int id, Channel channel, WebSocketSession session) {
+    public Client(Long id, Channel channel, WebSocketSession session) {
         this.id = id;
         this.channel = channel;
         this.session = session;
